--- Daycoval.Selenium.Framework/Parametros.cs
@@ -1,11 +1,0 @@
-﻿namespace Daycoval.Selenium.Framework
-{
-    public static class Parametros
-    {
-        public static string Login = "";
-        public static string Senha = "";
-        public static string DirImagens = "./images";
-        public static string NomeAqruivoBotao= "BotaoSenha{0}.jpg";
-
-    }
-}
