+++ Daycoval.Selenium.Framework/Parametros/Sistema.cs
@@ -1,0 +1,11 @@
+﻿namespace Daycoval.Selenium.Framework.Parametros
+{
+    public static class Sistema
+    {
+        public static string Login = "";
+        public static string Senha = "";
+        public static string NomeAqruivoBotao = "BotaoSenha{0}.jpg";
+        public static string DirTessData = "";
+        public static string DirImagens = "";
+    }
+}
