--- Daycoval.Selenium.Framework/Paramteros/Sistema.cs
@@ -1,10 +1,0 @@
-﻿namespace Daycoval.Selenium.Framework.Parametros
-{
-    public static class Sistema
-    {
-        public static string Login = "";
-        public static string Senha = "";
-        public static string DirImagens = "./images";
-        public static string NomeAqruivoBotao = "BotaoSenha{0}.jpg";
-    }
-}
